Fixed warning and cleanup
- Fixed DateBox test with locale
- Fixed timer and canvas warnings
- Moved mock data into utilities/test-utilities
