+++ apps/frontend/package.json
@@ -36,6 +36,7 @@
     "@types/react": "^18.2.61",
     "@types/react-dom": "^18.2.19",
     "axios-mock-adapter": "^1.22.0",
+    "canvas": "^2.11.2",
     "ts-jest": "^29.1.2",
     "typescript": "^5.4.2"
   },
