+++ apps/frontend/src/components/cln/CLNOffer/CLNOffer.test.tsx
@@ -1,101 +1,13 @@
 import { screen } from '@testing-library/react';
 import CLNOffer from './CLNOffer';
-import { ApplicationModes, Units } from '../../../utilities/constants';
-import { ListOffers } from '../../../types/lightning-wallet.type';
-import { renderWithMockContext } from '../../../utilities/test-utilities';
-
-export const mockListOffers: ListOffers = {
-  isLoading: false,
-  offers: [
-    {
-      bolt12: "lno1pwruza4cv00xgc8z2080w7kqjz5vunxrv4xvwu7fae60n0d4xqkdejwjcghwnghwhd0c8zugtfvslgzpfscghet0sqqqyqqqyqqqzqqqfqqqqrr9umvwg5a5nlkyjf0gz7wrc5gg40t8xflq2c23hyz0cyftmnz5kppafku2eurjg5rzr92kdwhd42w5pufheslp73jjl9euvvs32halutnggczp0kytwtrv28z2aw35z95m76x4ua2ds4l3nu05wt63u86jksql6t6qf",
-      offer_id: "Offer1",
-      active: true,
-      single_use: false,
-      used: false,
-      label: "MockOfferLabel"
-    }
-  ],
-  error: null
-};
+import { renderWithMockContext, getMockStoreData, mockFirstOffer } from '../../../utilities/test-utilities';
 
 describe('CLNOffer component ', () => {
-  const mockFaDollarSign = { icon: ['fas', 'dollar-sign'], iconName: 'dollar-sign', prefix: 'fas' };
   let providerProps;
-  
-  beforeEach(
-    () =>
-    (providerProps = {
-      authStatus: {
-        isAuthenticated: true,
-        isValidPassword: true
-      },
-      appConfig: {
-        appMode: { isLoading: true, unit: Units.SATS, fiatUnit: 'USD', appMode: ApplicationModes.DARK }
-      },
-      fiatConfig: { isLoading: true, symbol: mockFaDollarSign, venue: '', rate: 1 },
-      showToast: {
-        show: false,      
-        type: '',          
-        delay: 3000
-      },
-      walletBalances: {
-        isLoading: false,
-        clnLocalBalance: 0,
-        clnRemoteBalance: 0,
-        clnPendingBalance: 0,
-        clnInactiveBalance: 0,
-        btcSpendableBalance: 13.377777,
-        btcReservedBalance: 0
-      },
-      listOffers: mockListOffers,
-      listBitcoinTransactions: {
-        btcTransactions: [
-          {
-            account: 'mockAccount1',
-            type: 'onchain_fee',
-            credit_msat: '1000',
-            debit_msat: '2000',
-            currency: 'BTC',
-            timestamp: 1624000000,
-            tag: 'deposit',
-            txid: 'mockTxId',
-            blockheight: 123456,
-            outpoint: 'mockOutpoint',
-            origin: 'mockOrigin',
-            payment_id: 'mockPaymentId',
-            description: 'mockDescription',
-            fees_msat: '3000',
-            is_rebalance: false,
-            part_id: 123
-          },
-          {
-            account: 'mockAccount2',
-            type: 'onchain_fee',
-            credit_msat: '2000',
-            debit_msat: '3000',
-            currency: 'BTC',
-            timestamp: 1724000000,
-            tag: 'withdrawal',
-            txid: 'mockTxId2',
-            blockheight: 123900,
-            outpoint: 'mockOutpoint',
-            origin: 'mockOrigin2',
-            payment_id: 'mockPaymentId2',
-            description: 'mockDescription2',
-            fees_msat: '3000',
-            is_rebalance: false,
-            part_id: 124
-          }
-        ],
-        isLoading: false,
-        error: null
-      }
-    })
-  );
+  beforeEach(() => providerProps = JSON.parse(JSON.stringify(getMockStoreData())));
 
   it('should be in the document', () => {
-    renderWithMockContext(<CLNOffer offer={mockListOffers.offers![0]}/>, { providerProps });
+    renderWithMockContext(<CLNOffer offer={mockFirstOffer}/>, { providerProps });
     expect(screen.getByTestId('cln-offer-detail')).toBeInTheDocument();
     expect(screen.getByText('lno1pwruza4cv00xgc8z2080w7kqjz5vunxrv4xvwu7fae60n0d4xqkdejwjcghwnghwhd0c8zugtfvslgzpfscghet0sqqqyqqqyqqqzqqqfqqqqrr9umvwg5a5nlkyjf0gz7wrc5gg40t8xflq2c23hyz0cyftmnz5kppafku2eurjg5rzr92kdwhd42w5pufheslp73jjl9euvvs32halutnggczp0kytwtrv28z2aw35z95m76x4ua2ds4l3nu05wt63u86jksql6t6qf')).toBeInTheDocument();
   });
