+++ apps/frontend/src/components/shared/CurrencyBox/CurrencyBox.tsx
@@ -8,7 +8,7 @@ import Tooltip from 'react-bootstrap/Tooltip';
 
 import { AppContext } from '../../../store/AppContext';
 import { formatCurrency } from '../../../utilities/data-formatters';
-import { COUNTUP_DURATION, Units } from '../../../utilities/constants';
+import { APP_ANIMATION_DURATION, COUNTUP_DURATION, Units } from '../../../utilities/constants';
 import FiatBox from '../FiatBox/FiatBox';
 
 const CurrencyBox = props => {
@@ -24,7 +24,7 @@ const CurrencyBox = props => {
     const animation = animate(count, +formatCurrency(props.value, Units.SATS, appCtx.appConfig.unit, false, 5, 'number'), { duration: COUNTUP_DURATION });
     setTimeout(() => {
       setAnimationFinished(1);
-    }, 2000);
+    }, APP_ANIMATION_DURATION * 1000);
     return animation.stop;
     // eslint-disable-next-line react-hooks/exhaustive-deps
   }, []);
