+++ apps/frontend/src/components/shared/DateBox/DateBox.tsx
@@ -3,15 +3,9 @@ import React from 'react';
 import './DateBox.scss';
 import OverlayTrigger from 'react-bootstrap/OverlayTrigger'
 import Tooltip from 'react-bootstrap/Tooltip';
-
-const MONTH_NAMES = ['Jan', 'Feb', 'Mar', 'Apr', 'May', 'Jun', 'Jul', 'Aug', 'Sep', 'Oct', 'Nov', 'Dec'];
+import { convertIntoDateFormat } from '../../../utilities/data-formatters';
 
 const DateBox = props => {
-
-  const convertIntoDateFormat = () => {
-    let newDate = new Date(props.dataValue * 1000);
-    return newDate.getDate().toString().padStart(2, '0') + ' ' + MONTH_NAMES[newDate.getMonth()] + ', ' + newDate.getHours().toString().padStart(2, '0') + ':' + newDate.getMinutes().toString().padStart(2, '0');
-  }
 
   return (
     <OverlayTrigger
@@ -20,7 +14,7 @@ const DateBox = props => {
       overlay={(props.showTooltip ? <Tooltip data-testid='tooltip'>{props.dataType}</Tooltip> : <></>)}
       >
       <div className={props.className} data-testid='overlay-trigger'>
-        {convertIntoDateFormat()}
+        {convertIntoDateFormat(props.dataValue)}
       </div>
     </OverlayTrigger>
   );
